requirements.txt
pytorch-lightning
trimesh
onnxruntime
Pillow==9.0.0
pytorch
pymeshlab
scipy